fix: refactor project offset handling to utilize store for consistent positioning
